Move X_nodes and connectors to new homes
