+++ link-grammar/dict-common/build-disjuncts.c
@@ -14,13 +14,14 @@
 /* stuff for transforming a dictionary entry into a disjunct list */
 
 #include <math.h>
+#include "api-structures.h"
 #include "build-disjuncts.h"
 #include "dict-api.h"
 #include "dict-common.h"
 #include "dict-utils.h"
 #include "disjunct-utils.h"
-#include "externs.h"
 #include "string-set.h"
+#include "structures.h" // for SUBSCRIPT_MARK
 #include "tokenize/tok-structures.h" // XXX TODO provide gword acces methods!
 #include "utilities.h" /* For Win32 compatibility features */
 #include "word-utils.h"
