+++ link-grammar/dict-file/read-regex.c
@@ -11,8 +11,8 @@
 #include <string.h>
 #include "link-includes.h"
 #include "api-structures.h"
+#include "dict-common/dict-structures.h"
 #include "dict-common/file-utils.h"
-#include "structures.h"
 #include "read-regex.h"
 
 /*
