+++ link-grammar/structures.h
@@ -18,7 +18,6 @@
 
 #include "api-types.h"
 #include "api-structures.h"
-#include "dict-common/dict-structures.h"  /* For Exp, Exp_list */
 #include "histogram.h"  /* Count_bin */
 
 /* The following define the names of the special strings in the dictionary. */
@@ -33,9 +32,6 @@
 
 #define MAX_LINE 500          /* maximum width of print area */
 
-#define UNLIMITED_LEN 255
-#define SHORT_LEN 6
-
 /* Word subscripts come after the subscript mark (ASCII ETX)
  * In the dictionary, a dot is used; but that dot interferes with dots
  * in the input stream, and so we convert dictionary dots into the
@@ -44,58 +40,7 @@
 #define SUBSCRIPT_MARK '\3'
 #define SUBSCRIPT_DOT '.'
 
-typedef struct gword_set gword_set;
-
-/* On a 64-bit machine, this struct should be exactly 4*8=32 bytes long.
- * Lets try to keep it that way.
- */
-struct Connector_struct
-{
-	int16_t hash;
-	uint8_t length_limit;
-	             /* If this is a length limited connector, this
-	                gives the limit of the length of the link
-	                that can be used on this connector.  Since
-	                this is strictly a function of the connector
-	                name, efficiency is the only reason to store
-	                this.  If no limit, the value is set to 255. */
-	uint8_t nearest_word;
-	             /* The nearest word to my left (or right) that
-	                this could ever connect to.  Computed by
-	                setup_connectors() */
-	bool multi;  /* TRUE if this is a multi-connector */
-	uint8_t lc_start;     /* lc start position (or 0) - for match speedup. */
-	uint8_t uc_length;    /* uc part length - for match speedup. */
-	uint8_t uc_start;     /* uc start position - for match speedup. */
-	Connector * next;
-	const char * string; /* The connector name w/o the direction mark, e.g. AB */
-
-	/* Hash table next pointer, used only during pruning. */
-	union
-	{
-		Connector * tableNext;
-		const gword_set *originating_gword;
-	};
-};
-
-static inline void connector_set_string(Connector *c, const char *s)
-{
-	c->string = s;
-	c->hash = -1;
-}
-static inline const char * connector_get_string(Connector *c)
-{
-	return c->string;
-}
-
 typedef struct X_node_struct X_node;
-struct X_node_struct
-{
-	const char * string;       /* the word itself */
-	Exp * exp;
-	X_node *next;
-	const Gword *word;         /* originating Wordgraph word */
-};
 
 /**
  * Word, as represented shortly after tokenization, but before parsing.
